+++ NEWS.md
@@ -4,16 +4,25 @@ NEWS
 
 <!-- NEWS.md is generated from NEWS.Rmd. Please edit that file -->
 
-    #> Last Update: 2024-04-23 13:14:17.417653
+    #> Last Update: 2024-04-25 14:49:08.214479
 
 # Version History
+
+## Changes in version 1.0.2.9013
+
+- refactor: Modify scoring regime (AVERAGE_100_M10)
+  - Round scaling to 2 decimals
+  - Modify MetricScoring.xlsx
+  - `metric.scores`
+- refactor: Add new scoring regime (CAT_0510) to `metric.scores`
+- fix: Update MN_IBI_Fish thresholds, MetricScoring.xlsx
 
 ## Changes in version 1.0.2.9012
 
 - refactor: Update MetricScoring.xlsx
   - New scoring regime, AVERAGE_100_M10
   - Change MN_Bug_IBI scoring regime
-- refactor: Add new scoring regime calculation to `metric.values`
+- refactor: Add new scoring regime calculation to `metric.scores`
 
 ## Changes in version 1.0.2.9011